+++ libexec/bootstrap-scripts/main-dockerhub.sh
@@ -48,10 +48,49 @@ eval_abort "$SINGULARITY_libexecdir/singularity/bootstrap-scripts/pre.sh"
 eval_abort "$SINGULARITY_libexecdir/singularity/bootstrap-scripts/environment.sh"
 eval_abort "$SINGULARITY_libexecdir/singularity/python/import.py"
 
+# Docker layers are extracted with tar, but we need additional handling
+# for aufs whiteout files that may be present.
+# - .wh..wh..opq inside a directory indicates that directory is opaque.
+#    Any content in this directory from previous layers must be deleted
+#    before extracting the current layer.
+# - .wh.<file/dirname> indicates <file/dirname> from a previous layer must
+#    be deleted before we extract the current layer.
+# See:
+#   https://github.com/tonistiigi/docker/blob/2fb5d0c32376951ef41a6f64bb7dbd8f6fd14fba/pkg/archive/whiteouts.go#L3-L23
 for i in `cat "$SINGULARITY_CONTENTS"`; do
     name=`basename "$i"`
     message 1 "Exploding layer: $name\n"
-    zcat "$i" | (cd "$SINGULARITY_ROOTFS"; tar --exclude=dev/* -xf -) || exit $?
+    zcat "$i" | tar -tf - | grep '\.wh\.' | while read WHITEOUT
+    do
+        case "$WHITEOUT" in
+            # Handle opaque directories (remove them, will be recreated
+            # empty at extraction of tar).
+            */.wh..wh..opq )
+                OPAQUE_DIR=$(dirname "$WHITEOUT")
+                if [ -d "${SINGULARITY_ROOTFS}/${OPAQUE_DIR}" ]; then
+                    message 2 "Making $OPAQUE_DIR opaque\n"
+                    rm -rf "${SINGULARITY_ROOTFS}/${OPAQUE_DIR}"
+                fi
+                ;;
+            # Handle other plain whiteout marker files. Remove the target
+            # before extraction.
+            *)
+                WHITEOUT_TARGET=$(printf '%s' "$WHITEOUT" | sed -e 's@/.wh.@/@')
+                WHITEOUT_TARGET_ABS=$(readlink -f "${SINGULARITY_ROOTFS}/${WHITEOUT_TARGET}")
+                case "$WHITEOUT_TARGET_ABS" in
+                    "$SINGULARITY_ROOTFS"* )
+                        if [ -e "${WHITEOUT_TARGET_ABS}" ]; then
+                            message 2 "Removing whiteout-ed file/dir $WHITEOUT_TARGET\n"
+                            rm -rf "${WHITEOUT_TARGET_ABS}"
+                        fi
+                        ;;
+                esac
+                ;;
+        esac
+    done
+    # Now extract our current layer, exclude whiteout files handled
+    # above.
+    zcat "$i" | (cd "$SINGULARITY_ROOTFS"; tar --exclude=dev/* --exclude=*/.wh.* -xf - ) || exit $?
 done
 
 rm -f "$SINGULARITY_CONTENTS"